simplify metrics and logs impplementations
