+++ Steepfile
@@ -117,6 +117,7 @@ target :datadog do
   ignore 'lib/datadog/opentelemetry/sdk/id_generator.rb'
   ignore 'lib/datadog/opentelemetry/sdk/propagator.rb'
   ignore 'lib/datadog/opentelemetry/logs.rb' # references OpenTelemetry::SDK::Logs (external gem)
+  ignore 'lib/datadog/opentelemetry/signal_configuration.rb' # references OpenTelemetry::SDK::Resources (external gem)
   ignore 'lib/datadog/opentelemetry/sdk/logs_exporter.rb' # references OpenTelemetry::Exporter::OTLP::Logs (external gem)
   ignore 'lib/datadog/opentelemetry/sdk/span_processor.rb'
   ignore 'lib/datadog/opentelemetry/sdk/trace/span.rb'
