+++ lib/datadog/opentelemetry/logs.rb
@@ -1,20 +1,21 @@
 # frozen_string_literal: true
 
-require_relative '../core/configuration/ext'
-require_relative '../core/environment/socket'
+require_relative 'signal_configuration'
 
 module Datadog
   module OpenTelemetry
     class Logs
+      include SignalConfiguration
+
       EXPORTER_NONE = 'none'
 
       def self.initialize!(components)
         new(components).configure_logs_sdk
         true
       rescue => exc
-        components.logger.debug do
+        components.logger.warn(
           "Failed to initialize OpenTelemetry logs: #{exc.class}: #{exc.message}\n#{(exc.backtrace || []).join("\n")}"
-        end
+        )
         false
       end
 
@@ -34,42 +35,13 @@ module Datadog
         configure_log_record_processor(provider)
         ::OpenTelemetry.logger_provider = provider
 
-        # Avoid duplicate trace correlation fields when logs are also exported via OTel.
+        @logger.warn('OTel logs enabled: disabling Datadog log injection to prevent duplicate trace correlation fields')
         Datadog.configure do |c|
           c.tracing.log_injection = false
         end
       end
 
       private
-
-      def create_resource
-        resource_attributes = {}
-
-        @settings.tags&.each do |key, value|
-          otel_key = case key
-          when 'service' then 'service.name'
-          when 'env' then 'deployment.environment'
-          when 'version' then 'service.version'
-          else key
-          end
-          resource_attributes[otel_key] = value
-        end
-
-        resource_attributes['service.name'] = @settings.service_without_fallback || resource_attributes['service.name'] || Datadog::Core::Environment::Ext::FALLBACK_SERVICE_NAME
-        resource_attributes['deployment.environment'] = @settings.env if @settings.env
-        resource_attributes['service.version'] = @settings.version if @settings.version
-
-        hostname = Datadog::Core::Environment::Socket.resolved_hostname(@settings)
-        if hostname
-          if hostname == @settings.hostname
-            resource_attributes['host.name'] = hostname
-          elsif !resource_attributes.key?('host.name')
-            resource_attributes['host.name'] = hostname
-          end
-        end
-
-        ::OpenTelemetry::SDK::Resources::Resource.create(resource_attributes)
-      end
 
       def configure_log_record_processor(provider)
         exporter_name = @settings.opentelemetry.logs.exporter
@@ -81,6 +53,7 @@ module Datadog
       end
 
       def default_logs_endpoint(protocol)
+        # gRPC uses bare host:port (no path); HTTP uses /v1/logs per the OTLP spec.
         if protocol == 'grpc'
           "#{@agent_ssl ? 'https' : 'http'}://#{@agent_host}:4317"
         else
@@ -92,13 +65,14 @@ module Datadog
         require_relative 'sdk/logs_exporter'
 
         logs_config = @settings.opentelemetry.logs
-        protocol = get_logs_config_with_fallback(option_name: :protocol)
-        endpoint = get_logs_config_with_fallback(
+        protocol = config_with_fallback(signal: :logs, option_name: :protocol)
+        endpoint = config_with_fallback(
+          signal: :logs,
           option_name: :endpoint,
           computed_default: default_logs_endpoint(protocol)
         )
-        timeout = get_logs_config_with_fallback(option_name: :timeout_millis)
-        headers = get_logs_config_with_fallback(option_name: :headers)
+        timeout = config_with_fallback(signal: :logs, option_name: :timeout_millis) || 10_000
+        headers = config_with_fallback(signal: :logs, option_name: :headers)
 
         exporter = Datadog::OpenTelemetry::SDK::LogsExporter.new(
           protocol: protocol,
@@ -117,15 +91,6 @@ module Datadog
         provider.add_log_record_processor(processor)
       rescue LoadError => e
         @logger.warn("Could not load OTLP logs exporter: #{e.class}: #{e.message}")
-      end
-
-      # Returns logs config value if explicitly set, otherwise falls back to exporter config or computed default value.
-      def get_logs_config_with_fallback(option_name:, computed_default: nil)
-        if @settings.opentelemetry.logs.using_default?(option_name)
-          @settings.opentelemetry.exporter.public_send(option_name) || computed_default
-        else
-          @settings.opentelemetry.logs.public_send(option_name)
-        end
       end
     end
   end
