+++ lib/datadog/opentelemetry/metrics.rb
@@ -1,18 +1,20 @@
 # frozen_string_literal: true
 
 require_relative '../core/configuration/ext'
-require_relative '../core/environment/socket'
+require_relative 'signal_configuration'
 
 module Datadog
   module OpenTelemetry
     class Metrics
+      include SignalConfiguration
+
       EXPORTER_NONE = 'none'
 
       def self.initialize!(components)
         new(components).configure_metrics_sdk
         true
       rescue => exc
-        components.logger.error("Failed to initialize OpenTelemetry metrics: #{exc.class}: #{exc.message}: #{exc.backtrace.join("\n")}")
+        components.logger.warn("Failed to initialize OpenTelemetry metrics: #{exc.class}: #{exc.message}: #{exc.backtrace.join("\n")}")
         false
       end
 
@@ -41,35 +43,6 @@ module Datadog
 
       private
 
-      def create_resource
-        resource_attributes = {}
-
-        @settings.tags&.each do |key, value|
-          otel_key = case key
-          when 'service' then 'service.name'
-          when 'env' then 'deployment.environment'
-          when 'version' then 'service.version'
-          else key
-          end
-          resource_attributes[otel_key] = value
-        end
-
-        resource_attributes['service.name'] = @settings.service_without_fallback || resource_attributes['service.name'] || Datadog::Core::Environment::Ext::FALLBACK_SERVICE_NAME
-        resource_attributes['deployment.environment'] = @settings.env if @settings.env
-        resource_attributes['service.version'] = @settings.version if @settings.version
-
-        hostname = Datadog::Core::Environment::Socket.resolved_hostname(@settings)
-        if hostname
-          if hostname == @settings.hostname
-            resource_attributes['host.name'] = hostname
-          elsif !resource_attributes.key?('host.name')
-            resource_attributes['host.name'] = hostname
-          end
-        end
-
-        ::OpenTelemetry::SDK::Resources::Resource.create(resource_attributes)
-      end
-
       def configure_metric_reader(provider)
         exporter_name = @settings.opentelemetry.metrics.exporter
         return if exporter_name == EXPORTER_NONE
@@ -88,15 +61,16 @@ module Datadog
         require_relative 'sdk/metrics_exporter'
 
         metrics_config = @settings.opentelemetry.metrics
-        endpoint = get_metrics_config_with_fallback(
+        endpoint = config_with_fallback(
+          signal: :metrics,
           option_name: :endpoint,
           computed_default: default_metrics_endpoint
         )
-        timeout = get_metrics_config_with_fallback(option_name: :timeout_millis)
-        headers = get_metrics_config_with_fallback(option_name: :headers)
+        timeout = config_with_fallback(signal: :metrics, option_name: :timeout_millis)
+        headers = config_with_fallback(signal: :metrics, option_name: :headers)
         # OpenTelemetry SDK only supports http/protobuf protocol.
         # TODO: Add support for http/json and grpc.
-        # protocol = get_metrics_config_with_fallback(option_name: :protocol)
+        # protocol = config_with_fallback(signal: :metrics, option_name: :protocol)
         exporter = Datadog::OpenTelemetry::SDK::MetricsExporter.new(
           endpoint: endpoint,
           timeout: timeout / 1000.0,
@@ -111,15 +85,6 @@ module Datadog
         provider.add_metric_reader(reader)
       rescue LoadError => e
         @logger.warn("Could not load OTLP metrics exporter:  #{e.class}: #{e.message}")
-      end
-
-      # Returns metrics config value if explicitly set, otherwise falls back to exporter config or computed default value.
-      def get_metrics_config_with_fallback(option_name:, computed_default: nil)
-        if @settings.opentelemetry.metrics.using_default?(option_name)
-          @settings.opentelemetry.exporter.public_send(option_name) || computed_default
-        else
-          @settings.opentelemetry.metrics.public_send(option_name)
-        end
       end
     end
   end
