+++ sig/datadog/opentelemetry/logs.rbs
@@ -1,6 +1,8 @@
 module Datadog
   module OpenTelemetry
     class Logs
+      include SignalConfiguration
+
       EXPORTER_NONE: ::String
 
       @logger: Core::Logger
@@ -16,15 +18,11 @@ module Datadog
 
       private
 
-      def create_resource: () -> ::OpenTelemetry::SDK::Resources::Resource
-
       def configure_log_record_processor: (untyped provider) -> void
 
       def default_logs_endpoint: (::String protocol) -> ::String
 
       def configure_otlp_exporter: (untyped provider) -> void
-
-      def get_logs_config_with_fallback: (option_name: ::Symbol, ?computed_default: any) -> any
     end
   end
 end
