+++ sig/datadog/opentelemetry/metrics.rbs
@@ -1,6 +1,8 @@
 module Datadog
   module OpenTelemetry
     class Metrics
+      include SignalConfiguration
+
       EXPORTER_NONE: ::String
 
       @logger: Core::Logger
@@ -16,15 +18,11 @@ module Datadog
 
       private
 
-      def create_resource: () -> ::OpenTelemetry::SDK::Resources::Resource
-
       def configure_metric_reader: (::OpenTelemetry::SDK::Metrics::MeterProvider provider) -> void
 
       def default_metrics_endpoint: () -> ::String
 
       def configure_otlp_exporter: (::OpenTelemetry::SDK::Metrics::MeterProvider provider) -> void
-
-      def get_metrics_config_with_fallback: (option_name: ::Symbol, ?computed_default: any) -> any
     end
   end
 end
